# docs/swagger.yaml
basePath: /api
definitions:
  namespace.Namespace:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
        x-nullable: true
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  namespace.deleteRequest:
    properties:
      namespaceId:
        type: integer
      systemId:
        type: integer
    required:
    - namespaceId
    - systemId
    type: object
  namespace.getRequest:
    properties:
      id:
        type: integer
      name:
        type: string
      systemId:
        type: integer
    type: object
  namespace.getResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
        x-nullable: true
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  namespace.insertRequest:
    properties:
      name:
        type: string
      systemId:
        type: integer
    required:
    - name
    - systemId
    type: object
  namespace.insertResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
        x-nullable: true
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  namespace.listRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  namespace.updateRequest:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  namespace.updateResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
        x-nullable: true
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  route.deleteRequest:
    properties:
      id:
        type: integer
      namespaceId:
        type: integer
    required:
    - id
    - namespaceId
    type: object
  route.getByIdRequest:
    properties:
      id:
        type: integer
      namespaceId:
        type: integer
    required:
    - id
    - namespaceId
    type: object
  route.getResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
        x-nullable: true
      description:
        type: string
      id:
        type: integer
      method:
        type: string
      namespaceId:
        type: integer
      route:
        type: string
      tags:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    type: object
  route.insertRequest:
    properties:
      description:
        type: string
      method:
        type: string
      namespaceId:
        type: integer
      route:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - description
    - method
    - namespaceId
    - route
    type: object
  route.updateRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      method:
        type: string
      namespaceId:
        type: integer
      route:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - description
    - id
    - method
    - namespaceId
    - route
    type: object
...
  util.Error:
    properties:
      code:
        type: string
      message:
        type: string
      stack:
        type: string
    type: object
  util.ResponseWithError:
    properties:
      error:
        $ref: '#/definitions/util.Error'
    type: object
host: localhost:8081
info:
  contact:
    name: Sergey Korotkov
  description: Access control system
  title: BZDACS
  version: "1.0"
paths:
  /v1/namespace/create:
    post:
      consumes:
      - application/json
      description: Creates namespace
      operationId: namespace.create
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/namespace.insertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/namespace.insertResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - namespace
  /v1/namespace/delete:
    post:
      consumes:
      - application/json
      description: Deletes selected namespace
      operationId: namespace.delete
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/namespace.deleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - namespace
  /v1/namespace/get:
    post:
      consumes:
      - application/json
      description: Gets namespace with specified params
      operationId: namespace.get
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/namespace.getRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/namespace.getResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - namespace
  /v1/namespace/list:
    post:
      consumes:
      - application/json
      description: List of namespaces belong to selected system
      operationId: namespace.list
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/namespace.listRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/namespace.Namespace'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - namespace
  /v1/namespace/update:
    post:
      consumes:
      - application/json
      description: Updates namespace
      operationId: namespace.update
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/namespace.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/namespace.updateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - namespace
  /v1/route/create:
    post:
      consumes:
      - application/json
      description: Creates route
      operationId: route.create
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/route.insertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/route.getResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - route
  /v1/route/delete:
    post:
      consumes:
      - application/json
      description: Deletes route
      operationId: route.delete
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/route.deleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - route
  /v1/route/getById:
    post:
      consumes:
      - application/json
      description: Gets route by id
      operationId: route.getById
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/route.getByIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/route.getResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - route
  /v1/route/update:
    post:
      consumes:
      - application/json
      description: Updates route
      operationId: route.update
      parameters:
      - description: request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/route.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/route.getResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseWithError'
      tags:
      - route
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"